Rename files and fixed imports accordingly
